Add @types/bcrypt
