+++ src/routes/index.ts
@@ -1,2 +1,3 @@
 export * from './auth';
 export * from './tasks';
+export * from './users';
